// src/app/modules/student/student.interface.ts
import {  Model, Types } from 'mongoose';
import { IAcademicFaculty } from '../academicFaculty/academicFacultyInterface';
import { IAcademicDepartment } from '../academicDepartment/academicDepartment.interface';
import { IAcademicSemester } from '../academicSemester/academicSemester.interface';


export type UserName = {
    firstName: string;
    lastName: string;
    middleName: string;
}

export type Guardian = {
    fatherName: string;
    fatherOccupation: string;
    fatherContactNo: string;
    motherName: string;
    motherOccupation: string;
    motherContactNo: string;
    address: string;
}

export type localGuardian = {
    name: string;
    occupation: string;
    contactNo: string;
    address: string;
}

export type IStudent = {
    id: string;
    name: UserName;
    gender: 'male' | 'female';
    dateOfBirth: string;
    email: string;
    contactNo: string;
    emergencyContactNo: string;
    bloodGroup: 'A+' | 'A-' | 'B+' | 'B-' | 'AB+' | 'AB-' | 'O+' | 'O-';
    presentAddress: string;
    permanentAddress: string;
    guardian: Guardian;
    localGuardian: localGuardian;
    academicFaculty: Types.ObjectId | IAcademicFaculty;
    academicDepartment: Types.ObjectId | IAcademicDepartment;
    academicSemester: Types.ObjectId | IAcademicSemester;
    profileImage?: string;
}

export type StudentModel = Model<IStudent, Record<string, unknown>>